+++ src/main/scala/edu/ie3/simona/ontology/messages/flex/PowerLimitFlexOptions.scala
@@ -10,6 +10,7 @@ import edu.ie3.datamodel.models.result.system.{
   FlexOptionsResult,
   PowerLimitFlexOptionsResult,
 }
+import edu.ie3.simona.api.data.model.em.PowerLimitFlexOptions as ExtPowerLimitFlexOptions
 import edu.ie3.simona.exceptions.{CriticalFailureException, FlexException}
 import edu.ie3.simona.ontology.messages.flex.FlexibilityMessage.{
   IssueFlexControl,
@@ -19,6 +20,7 @@ import edu.ie3.simona.ontology.messages.flex.FlexibilityMessage.{
 import edu.ie3.simona.api.data.model.em.PowerLimitFlexOptions as ExtPowerLimitFlexOptions
 import edu.ie3.util.quantities.QuantityUtils.*
 import edu.ie3.util.scala.quantities.DefaultQuantities.*
+import edu.ie3.util.scala.quantities.QuantityConversionUtils.toSquants
 import squants.Power
 
 import java.time.ZonedDateTime
@@ -57,16 +59,6 @@ final case class PowerLimitFlexOptions(
       min + rhs.min,
       max + rhs.max,
     )
-
-  override def toExt(recipient: UUID, model: UUID): ExtPowerLimitFlexOptions = {
-    new ExtPowerLimitFlexOptions(
-      recipient,
-      model,
-      ref.toQuantity,
-      min.toQuantity,
-      max.toQuantity,
-    )
-  }
 }
 
 object PowerLimitFlexOptions extends FlexOptionsExtra[PowerLimitFlexOptions] {