+++ src/main/scala/edu/ie3/util/scala/quantities/QuantityUtil.scala
@@ -6,7 +6,6 @@
 
 package edu.ie3.util.scala.quantities
 
-import edu.ie3.simona.exceptions.QuantityException
 import squants.time.{Hours, TimeDerivative, TimeIntegral}
 import squants.{Quantity, Seconds, UnitOfMeasure}
 import tech.units.indriya.ComparableQuantity